+++ buisinessCase/src/main/java/com/humanbooster/buisinessCase/repository/LieuRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.humanbooster.buisinessCase.model.Lieu;
 
 @Repository
-public interface LieuRepository extends JpaRepository<Lieu, Integer>{
+public interface LieuRepository extends JpaRepository<Lieu, Long>{
 
 }
