+++ buisinessCase/src/main/java/com/humanbooster/buisinessCase/repository/VehiculeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.humanbooster.buisinessCase.model.Vehicule;
 
 @Repository
-public interface VehiculeRepository extends JpaRepository<Vehicule, Integer>{
+public interface VehiculeRepository extends JpaRepository<Vehicule, Long>{
 
 }
